Adding sound & music for plugins, added some sounds(destroyed, captured, select,cancel) +loading of maps from a directory.
Made units move faster in MoveAnimated so that they don't 'hop' to the next location. I hope this brings the illusion of smooth moving. If it is still not smooth we'll have to move units to the destination px/px.

Allowed Maps to be copied the idea is to load the originals once then take a copy of the map and use the copy. I'm copy constructor infected...
I had to do a weird cast on line 68 setTile((T) tileCopy); it would not take a Tile while I have the type T set to T extends Tile.