Fix Bug 28# CO causes unit to wait
So when you put a CO into a unit, the unit is still allowed to perform regular actions like move-> attack, capture, wait, etc

When you select the CO option, it'll cause your CO unit not to accept any more actions until the next turn. That is not correct, the CO unit can still perform actions after you load the CO unit into it as long as you didn't move the unit.